+++ src/ambition/7kaaInterface/inn.cc
@@ -1,7 +1,7 @@
 /*
  * Seven Kingdoms: Ambition
  *
- * Copyright 2025 Tim Sviridov
+ * Copyright 2025–26 Tim Sviridov
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -79,8 +79,8 @@ int getSelectedRecordNumber(
   return Ambition::getInnSelectedRecordNumber(inn, browserRecordNumber);
 }
 
-void guestLeft(
-  FirmInn* inn,
+void guestLeaving(
+  const FirmInn* inn,
   const VBrowseIF& guestBrowser,
   const int recordNumber
 ) {
@@ -92,7 +92,9 @@ void guestLeft(
     return;
   }
 
-  if (guestBrowser.rec_no == recordNumber) {
+  if (Ambition::getInnSelectedRecordNumber(inn, guestBrowser.rec_no)
+      == recordNumber
+  ) {
     guestSelectionValid = false;
   }
 }
@@ -124,7 +126,7 @@ void overrideHireButtonStatus(
   }
 }
 
-void refreshBrowser(
+bool refreshBrowser(
   const short firmRecordNumber,
   FirmInn* inn,
   VBrowseIF& browser,
@@ -132,7 +134,7 @@ void refreshBrowser(
   Button3D& hireButton
 ) {
   if (!Ambition::config.enhancementsAvailable()) {
-    return;
+    return false;
   }
 
   Ambition::refreshInnBrowser(
@@ -142,6 +144,8 @@ void refreshBrowser(
     removedUnitIndex,
     hireButton
   );
+
+  return true;
 }
 
 } // namespace _7kaaAmbitionInterface::Inn